roles(fix): preserve state for unknown reducer actions
